+++ cement/imports/ui/pages/exchangeRingPull-items.js
@@ -185,7 +185,7 @@ itemsTmpl.events({
     'keyup [name="qty"],[name="price"]': function (event, instance) {
         let qty = instance.$('[name="qty"]').val();
         let price = instance.$('[name="price"]').val();
-        qty = _.isEmpty(qty) ? 1 : parseInt(qty);
+        qty = _.isEmpty(qty) ? 1 : parseFloat(qty);
         price = _.isEmpty(price) ? 0 : parseFloat(price);
         let amount = qty * price;
         instance.state('amount', amount);
@@ -193,7 +193,7 @@ itemsTmpl.events({
     'click .js-add-item': function (event, instance) {
         let itemId = instance.$('[name="itemId"]').val();
         let qty = instance.$('[name="qty"]').val();
-        qty = qty == '' ? 1 : parseInt(qty);
+        qty = qty == '' ? 1 : parseFloat(qty);
         // let price = 0;
         //let amount = 0;
         let price = math.round(parseFloat(instance.$('[name="price"]').val()), 2);
@@ -226,7 +226,7 @@ itemsTmpl.events({
                     itemId: itemId
                 });
                 if (exist) {
-                    qty += parseInt(exist.qty);
+                    qty += parseFloat(exist.qty);
                     amount = math.round(qty * price, 2);
 
                     itemsCollection.update({
@@ -342,7 +342,7 @@ editItemsTmpl.events({
     'keyup [name="qty"],[name="price"]': function (event, instance) {
         let qty = instance.$('[name="qty"]').val();
         let price = instance.$('[name="price"]').val();
-        qty = _.isEmpty(qty) ? 0 : parseInt(qty);
+        qty = _.isEmpty(qty) ? 0 : parseFloat(qty);
         price = _.isEmpty(price) ? 0 : parseFloat(price);
         let amount = qty * price;
